+++ src/extension.ts
@@ -6,7 +6,7 @@ import { Lense, ok, isValidHex, Palette } from './fenlense';
 export function activate(context: vscode.ExtensionContext) {
     console.log('Congratulations, your extension "fen-lense" is now active!');
 
-    // "8/5k2/3p4/1p1Pp2p/pP2Pp1P/P4P1K/8/8 b - - 99 50"
+    // "8/5k2/3p4/1p1Pp2p/pP2Pp1P/P4P1K/8/8 b kKqQ e3 99 50"
 
     let def_light: string = "#f0d9b5";
     let def_dark: string = "#b58863";
@@ -36,7 +36,10 @@ export function activate(context: vscode.ExtensionContext) {
                 let filledBoard = await lense.populateBoard(boardPieces);
                 return new vscode.Hover(new vscode.MarkdownString(`![](data:image/png;base64,${filledBoard})`));
             } else {
-                console.log("failed to parse FEN String!")
+                console.log("failed to parse fen: " + boardPieces.message)
+                if (boardPieces.message[0] != '!') {
+                    return new vscode.Hover(new vscode.MarkdownString(`${boardPieces.message}`));
+                }
                 return null;
             }
         }