+++ CHANGELOG.md
@@ -9,3 +9,14 @@ All notable changes to the "fen-lense" extension will be documented in this file
 ## [0.1.1] - 2023.04.02
 
 - Add logo
+
+## [0.3.0] - 2023.09.03
+
+- Use new dependency 'Jimp' for rendering
+- Bundle dependencies with webpack
+
+## [0.4.0] - 2023.09.03
+
+- Allow fen strings surrounded in single quotes
+- Allow arbitrary amount of spaces between fields
+- Report parsing errors in hover